% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mditools-package.R
\docType{package}
\name{mditools-package}
\alias{mditools}
\alias{mditools-package}
\title{mditools: Microdata Infrastructure Tools for Firm-Level Microdata Research}
\description{
Supports the full analysis pipeline for researchers working with firm-level microdata. Provides data tools for panel preparation (import, outlier detection, classification harmonization), analytical methods (production function estimation, capital stock measurement, markups, intensity measures, distributions, regression, clustering), and disclosure tools for tagging outputs with dominance and observation counts before aggregation and publication.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Secretariat-CompNet/mditools}
  \item Report bugs at \url{https://github.com/Secretariat-CompNet/mditools/issues}
}

}
\author{
\strong{Maintainer}: Johanna Weiss \email{johanna.weiss@iwhesl.onmicrosoft.com}

Authors:
\itemize{
  \item Johanna Weiss \email{johanna.weiss@iwhesl.onmicrosoft.com}
  \item Eric Bartelsman
  \item Daniele Aglio
  \item Mirja Halbig
  \item Marco Miorandi
  \item Alessandro Zona Mattioli
}

Other contributors:
\itemize{
  \item Jing Chen [contributor]
  \item Juli<U+00E1>n D<U+00ED>az-Acosta [contributor]
  \item Alberto Ferreira [contributor]
  \item Javier Miranda [contributor]
  \item Marcelo Piemonte Ribeiro [contributor]
  \item Reetuparna Vishwanath [contributor]
  \item Chengzi Yi [contributor]
}

}
